+++ pkg/search/ibst_test.go
@@ -1,4 +1,4 @@
-package trees
+package search
 
 import "testing"
